+++ packages/react-static/src/static/components/BodyWithMeta.js
@@ -1,21 +1,11 @@
 import React from 'react'
-import jsesc from 'jsesc'
 import { pathJoin, makePathAbsolute } from '../../utils'
-
-const generateRouteInformation = embeddedRouteInfo => ({
-  __html: `
-    window.__routeInfo = JSON.parse(${jsesc(JSON.stringify(embeddedRouteInfo), {
-      isScriptContext: true,
-      wrap: true,
-      json: true,
-    })});`,
-})
 
 // Not only do we pass react-helmet attributes and the app.js here, but
 // we also need to  hard code site props and route props into the page to
 // prevent flashing when react mounts onto the HTML.
 const makeBodyWithMeta = async state => {
-  const { head, route, embeddedRouteInfo, clientScripts = [] } = state
+  const { head, route, inlineScripts, clientScripts = [] } = state
 
   // This embeddedRouteInfo will be inlined into the HTML for this route.
   // It should only include the full props, not the partials.
@@ -26,7 +16,7 @@ const makeBodyWithMeta = async state => {
       {!route.redirect ? (
         <script
           type="text/javascript"
-          dangerouslySetInnerHTML={generateRouteInformation(embeddedRouteInfo)}
+          dangerouslySetInnerHTML={{ __html: inlineScripts.routeInfo.script }}
         />
       ) : null}
       {!route.redirect